\name{drawtext}

\alias{drawtext}

\concept{drawing}
\concept{text}

\title{ Drawing primitives on images }

\description{
  FIXME: needs to be written
}

\section{ Usage }{
  \preformatted{
    ## S4 methods for signatures img='Image', xy='list', labels='list'
    ##                           img='Image', xy='matrix', labels='character'
    drawtext(img, xy, labels, ...)
}}

\section{ Arguments }{ \describe{


  \item{\code{img}}{ FIXME }

  \item{\code{xy}}{ FIXME }

  \item{\code{labels}}{ FIXME }
}}

\section{ Value }{

  FIXME
}

\section{ Details }{

FIXME
}

\seealso{
  \code{
    \link{Image-class}
  }
}

\author{
  Oleg Sklyar <\email{osklyar@ebi.ac.uk}>, 2007
}

\examples{
}


\keyword{file}

