\name{EBImage-defunct}

\alias{EBImage-defunct}
\alias{blur}
\alias{drawtext}
\alias{drawfont}
\alias{getFeatures}
\alias{hullFeatures}
\alias{zernikeMoments}
\alias{edgeProfile}
\alias{edgeFeatures}
\alias{haralickFeatures}
\alias{haralickMatrix}
\alias{moments}
\alias{rmoments}
\alias{smoments}
\alias{cmoments}

\docType{package}

\title{EBImage defunct functions}

\description{
  These following functions are defunct and will be removed in the next
  Bioconductor release.
}

\usage{
blur(...)
drawtext(...)
drawfont(...)
getFeatures(...)
hullFeatures(...)
zernikeMoments(...)
edgeProfile(...)
edgeFeatures(...)
haralickFeatures(...)
haralickMatrix(...)
moments(...)
rmoments(...)
smoments(...)
cmoments(...)
}

\arguments{
  \item{...}{Defunct arguments.}
}

\keyword{package}
