\name{EBImage-deprecated}

\alias{EBImage-deprecated}
\alias{blur}
\alias{equalize}
\alias{drawtext}
\alias{drawfont}
\alias{getFeatures}
\alias{hullFeatures}
\alias{zernikeMoments}
\alias{edgeProfile}
\alias{edgeFeatures}
\alias{haralickFeatures}
\alias{haralickMatrix}
\alias{moments}
\alias{rmoments}
\alias{smoments}
\alias{cmoments}

\docType{package}

\title{EBImage deprecated functions}

\description{
  These following functions are deprecated and will be defunct in the next
  Bioconductor release.
}

\usage{
blur(...)
equalize(...)
drawtext(...)
drawfont(...)
getFeatures(...)
hullFeatures(...)
zernikeMoments(...)
edgeProfile(...)
edgeFeatures(...)
haralickFeatures(...)
haralickMatrix(...)
moments(...)
rmoments(...)
smoments(...)
cmoments(...)
}

\arguments{
\item{...}{Deprecated arguments.}
}

\keyword{package}
