%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  EBImage-package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{EBImage-package}
\alias{EBImage-package}
\alias{EBImage}
\docType{package}



 \title{Dummy package EBImage}

 \description{
 \preformatted{
 Package: EBImage
Version: 999.999.999
Date: 2009-09-06
Title: DUMMY DUMMY DUMMY
Author: Henrik Bengtsson <henrikb@braju.com>
Maintainer: Henrik Bengtsson <henrikb@braju.com>
Depends: R (>= 2.9.0)
Description: DO NOT INSTALL THIS PACKAGE.  THIS IS AN EMPTY PACKAGE WITH THE SAME NAME AS AN EXISTING BIOC PACKAGE. THE ONLY PURPOSE OF THIS PACKAGE IS TO PREVENT THE R-FORGE BUILD SERVERS TO HALT THE BUILDING OF AROMA.CORE AND AROMA.AFFYMETRIX.  THIS DUMMY PACKAGE WILL NOT BE NEEDED WHEN R-FORGE SETS THE ENVIRONMENT VARIABLE R_check_force_suggests=false
License: Private
LazyLoad: TRUE

 }
 }


\keyword{package}
