\name{floodFill}

\alias{floodFill}
\alias{floodFill,array-method}
\alias{fillHull}
\alias{fillHull,IndexedImage-method}

\title{FIXME: Filling matrix data with flood fill algorithm}

\description{
  FIXME
}

\usage{
  \S4method{floodFill}{array}(x, pt, col, tolerance=1e-3, ...)
  \S4method{fillHull}{IndexedImage}(x, ...)
}

\arguments{

  \item{x}{ FIXME }

  \item{pt}{ FIXME }

  \item{col}{ FIXME }

  \item{tolerance}{ FIXME }

  \item{...}{ Reserved. }

}

\value{
  itself modified
}

\details{
}

\seealso{
  \code{
    \linkS4class{Image}, \linkS4class{IndexedImage}  }
}


\author{
  Gregoire Pau, Oleg Sklyar; 2007
}

\examples{
}

\keyword{file}

