\name{internals}
\alias{internals}

\alias{[,Image,missing,missing-method}
\alias{[,Image,missing,numeric-method}
\alias{[,Image,numeric,missing-method}
\alias{[,Image,numeric,numeric-method}
\alias{[,Image,array,missing-method}
\alias{[,Image,logical,missing-method}
\alias{colorMode,Image-method}
\alias{colorMode<-,Image,numeric-method}
\alias{compression,Image-method}
\alias{compression<-,Image,character-method}
\alias{fileName,Image-method}
\alias{fileName<-,Image,character-method}
\alias{imageData,Image-method}
\alias{imageData<-,Image,matrix-method}
\alias{imageData<-,Image,array-method}
\alias{Ops,Image,Image-method}
\alias{Ops,Image,numeric-method}
\alias{Ops,numeric,Image-method}
\alias{resolution,Image-method}
\alias{resolution<-,Image,numeric-method}
\alias{x}

\title{ Internals }

\description{
   See \code{\linkS4class{Image}} for the description of these accessor methods.
}

\keyword{internal}
