\name{EBImage-defunct}

\alias{EBImage-defunct}
\alias{animate}
\alias{blur}
\alias{drawtext}
\alias{drawfont}
\alias{getFeatures}
\alias{getNumberOfFrames}
\alias{hullFeatures}
\alias{zernikeMoments}
\alias{edgeProfile}
\alias{edgeFeatures}
\alias{haralickFeatures}
\alias{haralickMatrix}
\alias{moments}
\alias{rmoments}
\alias{smoments}
\alias{cmoments}
\alias{dilateGreyScale}
\alias{erodeGreyScale}
\alias{openingGreyScale}
\alias{closingGreyScale}
\alias{whiteTopHatGreyScale}
\alias{blackTopHatGreyScale}
\alias{selfcomplementaryTopHatGreyScale}

\title{Defunct functions in package \sQuote{EBImage}}

\description{These functions are defunct and no longer available.}

\details{
  The following functions are defunct and no longer available; use
  the replacement indicated below.
  
  \itemize{
    \item{\code{animate}: \code{\link{display}}}
    \item{\code{blur}: \code{\link{gblur}}}
    \item{\code{drawtext}: see package vignette for documentation on how to add text labels to images}
    \item{\code{drawfont}: see package vignette for documentation on how to add text labels to images}
    \item{\code{getFeatures}: \code{\link{computeFeatures}}}
    \item{\code{getNumberOfFrames}: \code{\link{numberOfFrames}}}
    \item{\code{hullFeatures}: \code{\link{computeFeatures.shape}}}
    \item{\code{zernikeMoments}: \code{\link{computeFeatures}}}
    \item{\code{edgeProfile}: \code{\link{computeFeatures}}}
    \item{\code{edgeFeatures}: \code{\link{computeFeatures.shape}}}
    \item{\code{haralickFeatures}: \code{\link{computeFeatures}}}
    \item{\code{haralickMatrix}: \code{\link{computeFeatures}}}
    \item{\code{moments}: \code{\link{computeFeatures.moment}}}
    \item{\code{cmoments}: \code{\link{computeFeatures.moment}}}
    \item{\code{rmoments}: \code{\link{computeFeatures.moment}}}
    \item{\code{smoments}: \code{\link{computeFeatures.moment}}}
    \item{\code{dilateGreyScale}: \code{\link{dilate}}}
    \item{\code{erodeGreyScale}: \code{\link{erode}}}
    \item{\code{openingGreyScale}: \code{\link{opening}}}
    \item{\code{closingGreyScale}: \code{\link{closing}}}
    \item{\code{whiteTopHatGreyScale}: \code{\link{whiteTopHat}}}
    \item{\code{blackTopHatGreyScale}: \code{\link{blackTopHat}}}
    \item{\code{selfcomplementaryTopHatGreyScale}: \code{\link{selfComplementaryTopHat}}}
  }
}

\keyword{package}
