\name{EBImage-deprecated}

\alias{EBImage-deprecated}
\alias{dilateGreyScale}
\alias{erodeGreyScale}
\alias{openingGreyScale}
\alias{closingGreyScale}
\alias{whiteTopHatGreyScale}
\alias{blackTopHatGreyScale}
\alias{selfcomplementaryTopHatGreyScale}

\title{Deprecated functions in package \sQuote{EBImage}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{EBImage} only, and will be defunct at the next release.
}

\details{
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{\code{dilateGreyScale}: \code{\link{dilate}}}
    \item{\code{erodeGreyScale}: \code{\link{erode}}}
    \item{\code{openingGreyScale}: \code{\link{opening}}}
    \item{\code{closingGreyScale}: \code{\link{closing}}}
    \item{\code{whiteTopHatGreyScale}: \code{\link{whiteTopHat}}}
    \item{\code{blackTopHatGreyScale}: \code{\link{blackTopHat}}}
    \item{\code{selfcomplementaryTopHatGreyScale}: \code{\link{selfComplementaryTopHat}}}
  }
}

\keyword{package}
