\name{EBImage-deprecated}

\alias{EBImage-deprecated}
\alias{read.image}
\alias{write.image}
\alias{edge.features}
\alias{edge.profile}
\alias{haralick.features}
\alias{haralick.matrix}
\alias{hull.features}
\alias{zernike.moments}
\alias{choose.image}
\alias{header}
\alias{assert}
\alias{matchObjects}
\alias{morphKern}
\alias{mkball}
\alias{mkbox}
\alias{frameDist}
\alias{stopIfNotImage}
\alias{chooseImage}

\alias{affinet}
\alias{resample}
\docType{package}

\title{EBImage deprecated functions}

\description{
  These functions are deprecated and may be defunct in a next Bioconductor release. 
}

\keyword{package}
