\name{EBImage-deprecated}

\alias{EBImage-deprecated}
\alias{choose.image}
\alias{read.image}
\alias{write.image}
\alias{write.image,Image-method}
\alias{edge.features}
\alias{edge.features,IndexedImage-method}
\alias{edge.profile}
\alias{edge.profile,IndexedImage-method}
\alias{haralick.features}
\alias{haralick.features,IndexedImage,Image-method}
\alias{haralick.matrix}
\alias{haralick.matrix,IndexedImage,Image-method}
\alias{hull.features}
\alias{hull.features,IndexedImage-method}
\alias{zernike.moments}
\alias{zernike.moments,IndexedImage,Image-method}

\docType{package}

\title{EBImage deprecated functions}

\description{
  The functions listed below are deprecated and their new variants should be used
  instead. The functions are likely to be defunct with the next Bioconductor 
  release, 2.2. The functions are deprecated due to the naming issues (the use
  of a dot) and new variants carry the same arguments as the deprecated ones.
}

\section{Substitutions for the deprecated functions}{

\code{'choose.image' => \link{chooseImage}}

\code{'edge.features' => \link{edgeFeatures}}

\code{'edge.profile' => \link{edgeProfile}}

\code{'haralick.features' => \link{haralickFeatures}}

\code{'haralick.matrix' => \link{haralickMatrix}}

\code{'hull.features' => \link{hullFeatures}}

\code{'read.image' => \link{readImage}}

\code{'write.image' => \link{writeImage}}

\code{'zernike.moments' => \link{zernikeMoments}}
  
}


\keyword{package}
