\name{EBImage-deprecated}

\alias{EBImage-deprecated}

\alias{frameDist}
\alias{matchObjects}
\alias{stopIfNotImage}
\alias{morphKern}
\alias{mkball}
\alias{mkbox}
\alias{header}
\alias{assert}
\alias{chooseImage}
\alias{resample}
\alias{sharpen}
\alias{umask}
\alias{modulate}
\alias{negate}
\alias{affinet}
\alias{normalize2}
\alias{noise}
\alias{mediansmooth}
\alias{cgamma}
\alias{enhance}
\alias{denoise}
\alias{contrast}
\alias{despeckle}
\alias{edge}
\alias{segment}
\alias{cthresh}
\alias{athresh}
\alias{channelMix}

\docType{package}

\title{EBImage deprecated functions}

\description{
  These following functions are deprecated and will be defunct in the next
  Bioconductor release.
   \itemize{
    \item \code{frameDist}, \code{matchObjects}
    \item \code{stopIfNotImage}, \code{morphKern}
    \item \code{mkbox}, \code{mkball}
    \item \code{header}, \code{assert}
    \item \code{chooseImage}
    \item \code{resample}, \code{sharpen}, \code{umask}, \code{modulate}
    \item \code{negate}, \code{affinet}, \code{normalize2}, \code{noise}
    \item \code{mediansmooth}, \code{cgamma}, \code{enhance}, \code{denoise}
    \item \code{contrast}, \code{despeckle}, \code{edge}, \code{segment} 
    \item \code{cthresh}, \code{athresh}
    \item \code{channelMix}
  } 
}

\keyword{package}
